@@ -29,12 +29,10 @@ xtb = ["xtb"]
 all = ["qcelemental", "qcengine", "geometric", "xtb", "tcpb"]
 
 [tool.poetry.group.dev.dependencies]
-black = "^24.0.0"
+ruff = "^0.5.0"
 mypy = "^1.3.0"
 pytest = "^7.3.2"
 pytest-cov = "^4.1.0"
-ruff = "^0.0.272"
-isort = "^5.12.0"
 pytest-mock = "^3.11.1"
 geometric = "^1.0.1"
 qcelemental = "^0.26.0"
@@ -55,20 +53,14 @@ mkdocs-exclude = "^1.0.2"
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.black]
-line-length = 88
-
-[tool.isort]
-force_grid_wrap = 0
-include_trailing_comma = true
-line_length = 88
-multi_line_output = 3
-use_parentheses = true
-
 [tool.ruff]
 line-length = 88
+target-version = "py39"
+
+
+[tool.ruff.lint]
 isort = { known-first-party = ["qcop", "tests"] }
-target-version = "py38"
+select = ["I", "F401"]
 
 [tool.coverage.run]
 branch = true
