@@ -6,7 +6,7 @@ https://github.com/grimme-lab/xtb-python/blob/main/xtb/qcschema/harness.py
 
 import importlib
 import os
-from typing import Callable, Optional, Tuple
+from typing import Callable, Optional
 
 import numpy as np
 from qcio import CalcType, ProgramInput, SinglePointResults, Wavefunction
@@ -102,7 +102,7 @@ class XTBAdapter(ProgramAdapter[ProgramInput, SinglePointResults]):
         update_func: Optional[Callable] = None,
         update_interval: Optional[float] = None,
         **kwargs,
-    ) -> Tuple[SinglePointResults, str]:
+    ) -> tuple[SinglePointResults, str]:
         """Execute xtb on the given input.
 
         Args:
